Refactor, change BaseObjectDetector class definition, make SSD model inherit from it
